+++ app/src/main/java/com/nweninge/reflex/SingleUserRecord.java
@@ -4,7 +4,7 @@ package com.nweninge.reflex;
  * Created by nweninge on 9/26/15.
  */
 public class SingleUserRecord extends Record {
-    public SingleUserRecord(double delayTime, double pressTime) {
+    public SingleUserRecord(long delayTime, long pressTime) {
         super(delayTime, pressTime);
     }
 }
